Updated for the new watch-only address support in mainline bitcoind.
WARNING: As of this commit, the old sipa watchonly build is NO LONGER SUPPORTED
due to wallet incompatibility.

With the next release of Bitcoin Core, MMGen will work with the binary build of
bitcoind and building the daemon from source will become unnecessary.  For those
who don't wish to wait, updated build instructions are provided in the
documentation wiki.