Stealth mnemonic entry
Different keyboard keys make different clicking sounds when pressed.  In
particular, the ENTER and SPACE keys are noticeably louder on most keyboards.
This makes possible an acoustic side-channel attack whereby an eavesdropper
with a microphone could gain knowledge about the lengths of the words in a
user's mnemonic seed phrase as they're being typed.  This information could
also be obtained by analyzing the pauses between words.

Since the words in the Electrum wordlist used by MMGen vary in length from three
to twelve letters, the keyspace required to be searched in a brute-force attack
on the seed could thus be drastically reduced.

Stealth mnemonic entry mitigates this vulnerability by treating all characters
that are not lowercase letters in the range 'a-z' as “pad characters”.  Pad
characters are simply discarded and can be typed either before, after or in the
middle of words.  For each word, once a sufficient number of characters,
including pad characters, have been typed (currently 12), any pad character
can be used to enter the word.  Since each word can now be entered using the
same number of keystrokes and padded with random characters at random
positions, acoustic analysis of the user's keyboard clicks will yield little
useful information for an attacker.

This feature is enabled by default as it doesn't impact existing functionality
and can simply be ignored if one chooses not to use it.