Version 0.11.0c - bugfix subrelease
- correctly handle privkey edge case
- correctly initialize altcoin trust levels

The edge case had a one in 2^127 chance of occurring per generated address,
so was *very* unlikely to have caused a problem for anyone.

The trust level bug caused an incorrect trust level to be displayed in altcoin
warning messages.  If you generate altcoin keys/addresses, an upgrade is
recommended.

MMGen's "trusted" altcoins BCH, LTC, ETH, ETC and XMR were not affected by the
bug.