proto.btc.tx.new: set sequence numbers for all inputs explicitly
Bitcoin Core v24.0.1 introduces a welcome but undocumented change in the
behavior of the ‘createrawtransaction’ RPC call: when one sequence number is
set, then the sequence numbers for the remaining inputs are automatically set
to the same value.  This occurs even when the ‘replaceable’ and ‘locktime’
arguments are not used.

To ensure this behavior across all coins and daemon versions, always set
sequence numbers for all inputs explicitly to the same value, ignoring
‘replaceable’ and ‘locktime’.