mmgen-{txcreate,txdo}: new `--autochg-ignore-labels` option
As of this commit, addresses in the tracking wallet with labels are considered
reserved, i.e. equivalent to used, for purposes of automatic change address
selection.

Use of this option restores the former default behavior of ignoring labels
during automatic change address selection, with one small difference: when
an address with a label is chosen as a change address candidate, a warning
is displayed before prompting the user for confirmation.

The option may also be set in the configuration file.