keygen.py: forbid use of non-safe public key generation backends
While the default generation backends for all public key types are timing
safe, a non-safe implementation could be selected manually by the user or
automatically as a fallback in the case of a broken installation (e.g.
missing libsecp256k1).

This patch prevents the use of non-timing-safe backends altogether except
when testing.